Add files via upload
Flashcards are an effective way to memorise new words. This project has been designed to help you learn the 100 most commonly used words used in French
Tkinter module has been extensively used to create the UI setup.  The window has been designed in such a way that it gives 5 seconds to recall the answer and then displays the corresponding answer.